feat : Lecture 삭제 API 추가

+++ src/main/java/com/teamexp/learnflowapi/lecture/service/LectureService.java
@@ -11,6 +11,7 @@ import com.teamexp.learnflowapi.lecture.dto.request.LessonCreateRequest;
 import com.teamexp.learnflowapi.lecture.dto.response.LectureFullCreateResponse;
 import com.teamexp.learnflowapi.lecture.dto.response.LectureResponse;
 import com.teamexp.learnflowapi.lecture.dto.response.PublishedResponse;
+import com.teamexp.learnflowapi.lecture.exception.LectureDeleteBlockedException;
 import com.teamexp.learnflowapi.lecture.exception.LectureNotFoundException;
 import com.teamexp.learnflowapi.lecture.exception.LectureInstructorUnauthorizedException;
 import com.teamexp.learnflowapi.lecture.model.*;
@@ -410,6 +411,10 @@ public class LectureService {
     @Transactional
     public void deleteLecture(Long lectureId, String instructorId) {
         Lecture lecture = findLectureWithValidation(lectureId, instructorId);
+        // lecture가 publish 상태이면 삭제 되면 안됨.
+        if (lecture.getStatus() == LectureStatus.AVAILABLE) {
+            throw new LectureDeleteBlockedException();
+        }
         lectureRepository.delete(lecture);
     }
